test: try creating new client
